Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -5,7 +5,7 @@ char* welcome() {
 }
 
 int main() {
-    double number; 
+    float number; 
     scanf("%4f", &number);
     printf("You entered: %4f\n", number);
     return 0;
